Show full URL for method
