+++ packages/zudoku/src/lib/plugins/openapi/OperationList.tsx
@@ -185,6 +185,7 @@ export const OperationList = () => {
               <StaggeredRender>
                 {tag.operations.map((fragment) => (
                   <OperationListItem
+                    serverUrl={result.data?.schema.url}
                     key={fragment.slug}
                     operationFragment={fragment}
                   />
